+++ Frontend/src/components/styles/feedpostskeleton.css
@@ -1,5 +1,5 @@
 .feed-post-skeleton {
-  width: 600px;
+  width: 800px;
   padding: 20px;
   border-radius: 8px;
   background-color: #fff;
